# perf: faster Function init