--- a/tinygrad/tensor.py
+++ b/tinygrad/tensor.py
@@ -14,10 +14,10 @@ from tinygrad.ops import LoadOps
 # An instantiation of the Function is the Context
 class Function:
   def __init__(self, device:str, *tensors:Tensor):
-    self.device = device
-    self.needs_input_grad = [t.requires_grad for t in tensors]
-    self.requires_grad = True if any(self.needs_input_grad) else None if None in self.needs_input_grad else False
-    if self.requires_grad: self.parents = tensors
+    self.device, self.parents, self.needs_input_grad, self.requires_grad = device, tensors, [tensors[0].requires_grad], tensors[0].requires_grad
+    for t in tensors[1:]:
+      self.needs_input_grad += [t.requires_grad]
+      self.requires_grad = True if True in self.needs_input_grad else None if None in self.needs_input_grad else False  # intentionally here, not called if len(tensors) = 1
 
   def forward(self, *args, **kwargs): raise NotImplementedError(f"forward not implemented for {type(self)}")
   def backward(self, *args, **kwargs): raise RuntimeError(f"backward not implemented for {type(self)}")