This is a statistical optimization, since the majority of functions get called with only one Tensor.

A similar optimization can be applied to `data` in the Tensor `__init__`.


before
```
codegen         mean runtime:  236.99ms, runs:   237.69,  297.91,  218.65,  222.19,  230.00,  223.47,  251.53,  227.39,  231.01,  230.01
methodcache     mean runtime:  228.42ms, runs:   221.12,  272.55,  206.80,  209.63,  209.35,  279.51,  234.99,  224.11,  212.05,  214.05
profile         mean runtime:  899.09ms, runs:   860.08,  904.68,  909.86,  877.98,  898.03,  891.36,  890.02,  879.26,  952.47,  927.21

Total time: 0.441247 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 16

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    16                                             @profile
    17                                             def __init__(self, device:str, *tensors:Tensor):
    18    173223      46488.4      0.3     10.5      self.device = device
    19    173223     215669.6      1.2     48.9      self.needs_input_grad = [t.requires_grad for t in tensors]
    20    173223     141448.4      0.8     32.1      self.requires_grad = True if any(self.needs_input_grad) else None if None in self.needs_input_grad else False
    21    173223      37640.2      0.2      8.5      if self.requires_grad: self.parents = tensors
```


after
```
codegen         mean runtime:  230.45ms, runs:   230.82,  282.35,  217.58,  215.76,  222.04,  217.84,  224.33,  243.53,  223.68,  226.56
methodcache     mean runtime:  217.25ms, runs:   197.53,  283.47,  206.02,  230.16,  206.30,  202.43,  216.83,  211.42,  204.91,  213.46
profile         mean runtime:  884.43ms, runs:   851.54,  890.50,  908.16,  877.49,  873.26,  857.97,  936.88,  875.90,  890.44,  882.20

Total time: 0.259016 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: __init__ at line 16

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    16                                             @profile
    17                                             def __init__(self, device:str, *tensors:Tensor):
    18    173223     140482.2      0.8     54.2      self.device, self.parents, self.needs_input_grad, self.requires_grad = device, tensors, [tensors[0].requires_grad], tensors[0].requires_grad
    19    173223      68926.6      0.4     26.6      for t in tensors[1:]:
    20     42921      24742.6      0.6      9.6        self.needs_input_grad += [t.requires_grad]
    21     42921      24864.0      0.6      9.6        self.requires_grad = True if True in self.needs_input_grad else None if None in self.needs_input_grad else False  # intentionally here, not called if len(tensors) = 1
````