@tinyb0t 2023-08-21
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/tensor.py                         4               4
------------------------------------------------------------
lines added in the tinygrad folder: 0
```