@wozeparrot 2023-08-21
I agree with @chaosagent here, I think these changes hurt readability a bit, and I don't think it's worth it for the perf, especially if it's a statistical optimization.